requirements.txt
fastapi
uvicorn
pydantic
dnspython
disposable-email-domains
